requirements.txt
openai==1.0.0
langchain==0.1.0
numpy==1.25.0
pandas==2.1.0
scikit-learn==1.3.0
pytest==8.2.0
pytest-mock==3.12.0
requests==2.32.0
tqdm==4.66.0
python-dotenv==1.1.1
rich==13.7.0
google-generativeai==0.5.2
matplotlib==3.8.2
fpdf2==2.7.9